+++ src/jsonq.h
@@ -1,17 +1,15 @@
 #ifndef JSONQ_H
 #define JSONQ_H
 
-#include <stdint.h>
-
 // JSON quick and dirty search for named value. Calls can be
 // concatenated to drill down through object layers.
 //
 // Accepts single quotes as a convenience.
 //
-// eg. char* s = "{'a':1,'b'2,'c':{'c1':1,'c1':2,'c3':3},'d',4}";
+// eg. char* s = "{'a':1,'b'2,'c':{'c1':1,'c2':2,'c3':3},'d',4}";
 //
 //   char tmp[256];
-//   int64_t v = jsonq(s,"c",tmp,sizeof(tmp))->jsonq_int(tmp,"c2");
+//   long long v = jsonq(s,"c",tmp,sizeof(tmp))->jsonq_int(tmp,"c2");
 //
 // Doesn't use any memory other than what's on the stack.
 //
@@ -19,7 +17,7 @@
 // and enumerate/index arrays directly.
 
 extern const char* jsonq(const char* s, const char* name, char* dstbuf, int dstlen);
-extern int64_t jsonq_int(const char* s, const char* name);
+extern long long jsonq_int(const char* s, const char* name);
 extern double jsonq_real(const char* s, const char* name);
 extern int jsonq_bool(const char* s, const char* name);
 extern int jsonq_null(const char* s, const char* name);
